==> 02.encoding.ipynb <==
## Encoding 

import pandas as pd
import numpy as np
import sklearn as sk
import matplotlib.pyplot as plt
import plotly.express as px
from sklearn.preprocessing import LabelEncoder, OneHotEncoder

train_without_nan = pd.read_csv("data/train_without_nans.csv")
test_without_nan = pd.read_csv("data/test_without_nans.csv")

print(train_without_nan.shape)
train_without_nan.info()

## Features' dtypes

First of all we need to understand if features' dtypes really corresponds to real features' meaning and if there is no mistake in dtype for a particular feature.

To ensure that all is OK we need to check data_description.txt file and ensure that dtypes really correspond to the features

Now let's separate object features from numerical ones

object_columns = train_without_nan.select_dtypes(include='object')
numerical_columns = train_without_nan.select_dtypes(exclude="object").drop(["SalePrice"], axis=1)

## Category features encoding

Let's check how many category features we have 

object_columns.shape

And and what is the sum of the counts of categories in each feature

category_counts_train = [train_without_nan[column].value_counts().shape[0] for column in object_columns.columns]
print(sum(category_counts_train))
print(category_counts_train)

category_counts_test = [test_without_nan[column].value_counts().shape[0] for column in object_columns.columns]
print(sum(category_counts_test))
print(category_counts_test)

We see that if we use, for example, one hot encoding, we will have additional 265 columns in our data (minus 43 because we will delete previous columns of taken feature).

Anyway we have no better choice than encode features which have many categories with one hot encoding. To make our dataset at least a little bit smaller we will use label encoding for features which have only 2 categories. 

encoded_train = train_without_nan.copy()
encoded_test = test_without_nan.copy()

for column in object_columns.columns:
    category_cnt = encoded_train[column].value_counts().shape[0]
    
    if category_cnt == 2:
        encoder = LabelEncoder()
        encoded_train[column] = encoder.fit_transform(encoded_train[column])
        encoded_test[column] = encoder.transform(encoded_test[column])
    else:
        encoder = OneHotEncoder(sparse_output=False)
        
        one_hot_encoded_train = encoder.fit_transform(encoded_train[[column]])
        one_hot_encoded_test = encoder.transform(encoded_test[[column]])
        
        one_hot_df_train = pd.DataFrame(one_hot_encoded_train,
                                        columns=encoder.get_feature_names_out([column]))
        one_hot_df_test = pd.DataFrame(one_hot_encoded_test,
                                        columns=encoder.get_feature_names_out([column]))
        
        encoded_train = pd.concat([encoded_train, one_hot_df_train], axis=1)
        encoded_train = encoded_train.drop([column], axis=1)

        encoded_test = pd.concat([encoded_test, one_hot_df_test], axis=1)
        encoded_test = encoded_test.drop([column], axis=1)

encoded_train.shape

encoded_test.shape

## Casting continuous features

To work with all features we need to convert them to float. Because of the fact that we have encoded all categorical features now we have all our features numerical

encoded_train = encoded_train.astype("float64")
encoded_test = encoded_test.astype("float64")

encoded_train.to_csv("data/encoded_train.csv", index=False)
encoded_test.to_csv("data/encoded_test.csv", index=False)
